+++ src/load_balanced/writhe.c
@@ -23,7 +23,7 @@
 int writhe(const struct link* L) 
 {
     int writhe = 0;
-    for (int component = 0; component < L->number_of_components; component++) {
+    for (size_t component = 0; component < L->number_of_components; component++) {
         if (L->number_of_crossings_in_components[component] == 0) {
             continue;
         }
